Add support for :parse-literal hook.
This hook takes a literal non-table AST node, a source table, and the raw
string which parsed to that node. Whatever non-nil value it returns
will be placed into the AST.

This can be used to work around the lack of source map data on
non-table AST nodes like strings and numbers, either by wrapping it in
another AST node, or by recording it elsewhere.