lambda doesn't check args that start with _
We decided in the usergroup on May 18, 2024 that we should treat
anything that starts with _ the same way as we treat _ in lambda. I
think this makes it more consistent with match/case patterns.

Also, I changed it from `match` to `find` for superstitious performance
reasons.

I imagine that str:match has to construct a string and str:find
doesn't, but I've never tested if it actually affects performance.