make macrodebug emit valid code on cyclic ASTs
When macrodebug expands to a cyclic/recursive AST like

  (macrodebug (case [1 2] (where (or [x y] [y nil x]) (= 3 (+ x y))) x))

it currently emits placeholders @1, @2, etc due to the implicit
`:detect-cycles? true` passed to fennel.view, resulting in invalid
fennel syntax in the serialized output.

This change simply disables detect-cycles. This fixes the case where
a macro expands to an AST with references to node higher in the tree
like the above, which covers most edge cases.

This makes it *possible* to emit an infinitely recursive AST, which with
this change would result in a massive expansion that eventually gives up
upon hotting the maximum depth for fennel.view, but this is a mutch
rarer edge case.

This should eventually be revisited to any failure cases involving
infiniten AST's serialization can give a useful compiler error, however.
Ideally, this will eventually be covered by @-macros such that @<n>
forms can actually be evaluated as fennel syntax back into a
self-referencing table.