tset add assert for bad targets
Previously, (tset 1 1 1) would compile to 1[1] = 1, which is not
syntactically valid lua. This change makes it a fennel compiler error
instead of a lua compiler error.
Also, (tset true 1 1) is now a compiler error.

It hurts my soul, but `(tset :foo 1 1)` now correctly compiles to
("foo")[1] = 1, with the parentheses wrapping the string.