remove IIFE's from (values ..) destructuring
NOTE: This commit builds on dbb10b59250a1dbcd2b7a86d0ef5ae0e538b0388,
which changed `values` to propagate nval. This commit is an example that
builds further on that to eliminate more IIFE's.

For the cases (local x (values (do 1))) or (pick-values 2 :x (do :y)),
we get a remnant IIFE generated by the `do`. As of commit
, `values` is now propagating
nval when compiling its operands, but this doesn't account for
assignment, where nval is known, but not propagated by
compiler.destructure. In this case, while we could make a wider change
to destructure, here we parse opts.target from `values` to resolve nval.

Resolving nval successfully removes the IIFE's, but as a side effect,
introduces order of execution edge cases - which may or may not
matter, since order of operations is not guaranteed here, but it's
surprising behavior when it changes the outcome.

--- Example 1:
See the last new test case introduced by this commit in `test-values` vs
the first in `pick-values`. Because there's no target, this code:

    (do (var i 0) (fn i++ [] (set i (+ 1 i)) i)
        [(values (i++) (values (do (i++))))])

compiles the final `(do (i++))` to an IIFE, so it outputs [1 2].
However, as a result of this patch, the same code with `pick-values`:

    (do (var i 0) (fn i++ [] (set i (+ 1 i)) i)
        [(pick-values 2 (i++) (values (do (i++))))])

eschews the IIFE, outputting [2 1]. It compiles to
    local i = 0
    local function _2b_2bi()
      i = (1 + i)
      return i
    end
    local pv_1_, pv_2_ = nil, nil
    local _3_
    do
      _3_ = _2b_2bi()
    end
    pv_1_, pv_2_ = _2b_2bi(), _3_
    return {pv_1_, pv_2_}

While order of execution isn't guaranteed, as one would still normally
expect the same result between `values` and `pick-values` when dealing
with a fixed number of values.

---- Example 2:

This one is likely less of an issue, but I mention it for completeness.

When `(values)` calls are nested, the result of the expression is still
the same in the absence of effects, but with side effects order of
execution is changed.

Nontrivial expressions (lists) passed to (values ...) beyond `nval` will
be treated as a side effect and pulled out to run *before* the
assignment/binding. Nesting `(values)` is unnecessary, but the compiler
*does* allow it. For a live example, see the changed test case from [1 2
4] to [2 3 4], which became necessary after this patch.

----

For example, given: `(pick-values 2 (i++) (i++) (values (i++) (i++)))`
or equivalent: `(local (x y) (values (i++) (i++) (values (i++) (i++))))`

The Lua output is in the form of:
    local pv_18_, pv_19_ = nil, nil
    _2b_2bi()
    pv_18_, pv_19_ = _2b_2bi(), _2b_2bi(), _2b_2bi()

The number of exprs "lifted" in this way depends on how many calls occur
inside the nested (values). For example:
    (local (x y) (values (values (i++) (i++) (i++) (i++))))
and:
    (local (x y) (values (i++) (values (i++) (i++) (i++))))
BOTH compile to something like:
    local x, y = nil, nil
    _2b_2bi()
    _2b_2bi()
    x, y = _2b_2bi(), _2b_2bi()

But a flat `(local (x y) (values (i++) (i++) (i++) (i++)))` emits
    local a1, b1 = _2b_2bi(), _2b_2bi(), _2b_2bi(), _2b_2bi()
as expected.