stop REPL tests from leaking _G.___replLocals___
Running the REPL on the global env in the test/repl.fnl suite was
leaking both _G.___replLocals___ and _G.___repl___ into the same global
env used by faith.

This was giving a false positive to faith's check against
_G.___replLocals___ to skip invokign (os.exit 1), resulting in every
test run not using PUC Lua 5.1 (which skips the REPL tests) to exit with
a zero status code even when test fail.

This change both saves off any existing values in a setup-all hook (just
in case the tests are somehow run from a REPL) and restores them
afterwards, as well as clearing them between test.