Rewrite parse-string to not use _G.load.
Currently only supports the escapes present in 5.1.

Using _G.load in the parser saves us having to write our own string
unescaping functionality, but it's a bit of a cheat. It's not ideal
because we really want to be able to use the parser inside sandboxes
which do not expose _G.load. It turns out that exposing _G.load in a
sandbox is a super bad idea, because it not only loads code from
strings, it also allows loading *bytecode* from strings, and the
bytecode is not checked for validity. So sandbox escapes using
malformed bytecode are fairly common, such as
https://memorycorruption.net/posts/rce-lua-factorio/

This implements all the escapes in the Fennel test suite. However,
Fennel strings are defined in terms of Lua strings, which means they
vary depending on which version of Lua you use. The test suite
currently only covers the decimal escape syntax from 5.1, but in later
versions they added hex escapes and unicode-style escapes, plus a \z
escape. These have not yet been implemented.