fix dynamic set support for globals + hashfn args
* Fixes `(set (_G :x :y) 1)`, `#(set $ :x $2)`, etc
* Improves error message when invoking dynamic set without keys

In the case of globals and hashfn args, a custom assertion on `(.
scope.manglings (tostring target))` was causing a compiler error. These
are already accounted for in symbol-to-expression, which
dynamic-set-target invokes, so we can simply omit the assertion. This
will help avoid bugs resulting from maintaining the same logic in more
than one place.