editorconfig: Comment markdown trailing whitespace
In markdown, two trailing spaces at EOL guarantees a line break in the
output.

For now, leaving a comment explaining the setting, but this is a
somewhat obscure markdown feature, and we may want to consider leaving
it turned off since, even if we exlpicitly allow it in editorconfig, a
contributor seeing trailing whitespace in the wild is most likely to
simply assume it was in error.