test/irc.lua: `git remote get-url` to check origin
When checking that the CI job is running in the official git.sr.ht
repository, use `git remote get-url origin` instead of of
`git remote show origin` so it only provides the URL, and ':read('*l')`
to restrict the output to the first line.

Switching to `:read('*l')` when getting the branch as well, rather than
stripping the newline with gsub.