Precompile macros when bootstrapping the compiler.
Going all the way to Lua code with the macros requires a few tweaks.

Currently we accept :_COMPILER in the :env field of the options table
to evaluate code in compiler scope. However, we also need a similar
option for scope in order to compile macros in eval-compiler. This has
to be added to both the bootstrap compiler and the mainline compiler.
This is an undocumented internal feature only.

We also disable global mangling in the bootstrap compiler and replace
it with _G lookups, because global mangling is dumb, and we don't have
backwards-compatibility concerns in the bootstrap compiler.

Currently we're disabling metadata for built-in macros because loading
metadata under 5.4 (but not 5.1 or luajit!!!) causes performance to
tank when we try to load the module to find a place to set the metadata.

We also need to avoid having `assert-repl' emit a call to `collect'
because the sequence metatables for the arglist definition vs the
check on the emitted code don't match. It's easy enough to replace it
with an `each' but this will probably trip us up in the future.

The tests are still failing because of metatable mismatches when
loading the compiler from the compiler. The `test-nest' test has been
disabled for now.