Add workaround for pandoc bug.
Older versions of pandoc try to set the font to a font that doesn't
actually exist. Older versions of groff don't care, but newer ones
will catch the mistake.

This has been fixed in pandoc 3 but we're not ready to declare a
dependency on a newer version since installing the whole haskell
toolchain is a lot for something that can be fixed with a trivial
sed.

https://github.com/jgm/pandoc/issues/9020