don't modify non-string and non-nil values in compiler.traceback
Lua's debug.traceback entry from the manual:

If message is present but is neither a string nor nil, this function
returns message without further processing. Otherwise, it returns a
string with a traceback of the call stack. The optional message string
is appended at the beginning of the traceback. An optional level
number tells at which level to start the traceback (default is 1, the
function calling traceback).