Allow disabling global mangling and using _G['what-ever'] instead.
Use this when bootstrapping the built-in macros to avoid issues with
double-loading the compiler.

It's unclear why, but when you load the compiler with itself, the
second load uses a scopes.compiler table that's unaware of built-ins
like `list?' and so mangles that into something that will definitely
not resolve to anything sensible.