add syntax for infinity and NaN
.inf and -.inf were added to represent positive and negative infinity.
.nan and -.nan were added to represent positive and negative NaN (not
a number) values.  For some reason, on x86 in PUC Lua 0/0 gives -nan,
but nan on ARM, so to generate positive NaN portably across most Lua
implementations, the nan is first converted to a string and is checked
to contain a minus sign.  LuaJIT and many other implementations do not
differentiate between NaN and negative NaN, so the tests only check
for positive NaN