Skip IRC reporting if we can't determine Git remote
This means either Git is not installed, or the test is being run from a
published tarball and not a Git checkout.

In either case, that means we don't want to report the build result
(many Linux distributions also build with network access disabled, so
reporting will fail anyway) - and we should exit with status code 0, as
this is not actually an error.

Signed-off-by: Michel Lind <salimma@fedoraproject.org>