Improve tests for repl readChunk newline behavior.
The fundamental problem here is that readChunk is a bad abstraction
for cases where you're not taking input from a stream. In most
non-shell cases, you don't really want to send partial chunks of input
one at a time; you know you have a complete form, and you want to send
it to the parser.

In this case if you use readChunk, you have to remember every time to
append whitespace to the end of the input you're sending; otherwise
the parser can't tell that the difference between "I've typed 15 and I
want you to evaluate it" vs "I've started to type 1500 but the
chunking just happened to break right in the middle".

I suspect we would be best off adding a new field to the options
table; read-form or something instead of framing it in terms of chunks.