Don't print sparse tables as sequences by default.
This causes a lot of confusion among people who don't already
understand the behavior of nils in sequences, because it implies that
a table that has gaps in it will still be treated like a
sequence. Once they go to use it with ipairs, they will be confused as
to why certain elements are ignored.

It would be better to print these as curly-bracket tables, because it
better reflects their true nature. The option to set :max-sparse-gap
if you want non-sequences to look sequency is still there.