merge passed repl view-opts with custom defaults
The REPL defaults/overrides are merged using `copy`, but the way it was
prevoiusly implemented meant that in the following code:

    (set fennel.repl.view-opts {:max-sparse-gap 5})
    (fennel.repl {:view-opts {}})

any defaults set on the REPL would be wiped out by the presence of an
empty :view-opts table.

It may be more maintainable in the long run to simply implement a
standard utils.merge-with instead of playing games with `copy`, but this
gets the job done in the meantime.