Add `pack` to macro environment
The `unpack` helper on its own lacks symmetry; `pack` in particular is
frequently useful in tandem with `unpack` when storing varargs or the
results of a function with multiple returns for later unpacking, and
I've often found myself quickly implementing it in macro modules.

Also adding `unpack` to the utils module caused a net drop in LOC due to
all of the modules with the `(local unpack (or _G.unpack table.unpack))`
boilerplate.

I also considered adding `maxn`, but this comes up less frequently, and
is really only needed for defensive coding against tables from unknown
sources, such as when writing a library/api. Since macros receive `nil`
as a symbol when invoked, it's less likely to come up when writing
macros, so left it out for now.