update from-clojure CljLib notices, and add deps.fnl and deps.edn to the mix
I'd like to propose a 'one-line-per-sentence' approach for handling shuch documents.
Because it is way easier to deal with them when viewing pathces, than when paragraphs are formatted.
Additionally, it helps to see way too long sentences easily and structure text better.

https://rhodesmill.org/brandon/2012/one-sentence-per-line/

I've been employing this for years, and as you can see I'm also using it in this commit message.
For real, works amazingly for text diffs.