Precompile macros to disk using aot.lua instead of in-process.
It would be cleaner to do this in-process like we used to, and it
works fine for normal use. But when it's used to make fennel-bin, the
precompiled-macros build of regular fennel tries to precompile the macros
again and use those precompiled macros to build other parts of the compiler,
and that fails with a nonsense error:

    lua/src/lua: src/fennel/binary.fnl:165:?: Compile error: unknown identifier: fennel

The filename and line number are wrong, (and not by just a few lines)
so it's nearly impossible to find out what's actually going on here.

Rather than try and untangle this, we can bypass the whole mess by
using the bootstrap compiler to precompile the macros, and putting
them in build/macros.lua and build/match.lua. I thought this would
just be a first step that would help me understand where the problem
was coming from, but instead it make the problem go away!