unload macro module before actually reloading the module
This way if the macro module has the same module name, as the runtime
module, i.e. when using init.fnl and init-macros.fnl in the same
directory, the macro module is unloaded first, then require will load
it back.  In case of error, the old state of the macro module is
re-set to the macro-loaded table.