A few fixes to squash the final linting warnings.
Don't use a custom env for macros! Since fennel-ls really doesn't like
this, pass in the extra things we need using ... instead.

Abort embed-src if _G.io isn't present. (a bit tacky)

One weird thing here is that utils is *also* used by bootstrap/aot.lua
in order to store the metadata, but that's not going to be understood
by fennel-ls so if we remove the calls to utils.copy or
utils.idempotent-expr then it would falsely lint as not used but it is
in fact used by metadata. Weird but ok.