Add AOT Bootstrap Compiler to `make fennel` Prerequisites
The `Makefile` rule the `fennel` script was missing a prerequisite to
the AOT bootstrap compiler (`bootstrap/aot.lua`). This prerequisite is
notably present in the `fennel.lua` library rule, and makes sense to
include in the script - changes to `bootstrap/aot.lua` would cause
changes to the compiled `fennel` script.