test.fuzz-string new test to check parse vs compile-string
We've found a couple of bugs involving string constants being
incorrectly serialized to Lua code. This change adds a new fuzz test
to test.fuzz-string to find that category of bugs.
The test asserts that randomly generated strings evaluate to themselves.
It will run on all versions of Lua, unlike the other string fuzz test,
which only works on Lua 5.3+ (for unicode escapes).