improve line-length hanling when printing associative tables
Before this change, associative tables that had only one key, and a
really long value would not be split into multiple lines despite the
:line-length setting in the view options:

>> {:foo "baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaar"}
{:foo "baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaar"}

After this change, a newline is properly included:

>> {:foo "baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaar"}
{:foo
 "baaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaar"}

This slightly changes how nested data structures are printed:

>> {:sequence ["fooooooooooooo" "baaaaaaaaaaaaar" "baaaaaaaaaaaaaz" "quuuuuuuuuuuuux"]}
{:sequence
 ["fooooooooooooo" "baaaaaaaaaaaaar" "baaaaaaaaaaaaaz" "quuuuuuuuuuuuux"]}

Here, the sequence itself doesn't pass the threshold, but the overall
map does.

Here, however, both the map and the sequence are above the threshold

>> {:sequence ["fooooooooooooooooo" "baaaaaaaaaaaaaaaaar" "baaaaaaaaaaaaaaaaaz" "quuuuuuuuuuuuuuuuux"]}
{:sequence
 ["fooooooooooooooooo"
  "baaaaaaaaaaaaaaaaar"
  "baaaaaaaaaaaaaaaaaz"
  "quuuuuuuuuuuuuuuuux"]}

This means there would be less indentation in the nested data
structures that was previously introduced by long keys

Added a bit more tests, and made tests more readable with a special
function to allow writing strings with more visually readable
indentation.