alphabetize and organize friend.fnl suggestions
The suggestions table in `friend.fnl` has gotten to the size where it's
challenging to format or to know where to add a new suggestion.

This change adds sugar that makes it possible to add suggestions
one-by-one, and alphabetizes the existing suggestions so it's clear
where to add existing suggestions.