fix: case shouldn't try to bind & symbols
Inside of (where (or [a & b] [a b & c])) style patterns, the rule is
that symbols that are bound in all of the subpatterns should be bound by
the (or) clause. However, this logic is separate from the actual binding
logic, and incorrectly decides that `&` is supposed to be bound, since
it is present in both patterns.

With this change, all `&` pseudo-symbol things won't count as a binding
for (or).