make changes to macro arglists based on feedback
update icollect to use a single argument list

This demonstrates that collect also supports a separate form for keys
and values in the docstring instead.

---

clarify that iterators are not always functions in macro arglists

Instead, `*iterator-values` is used to demonstrate that it's one or more
values; however, this isn't a widely used convention so this may not be
correct.

---

replace step? with ?step in for-loop binding macros