Don't let globals trigger locals-saving.
We're currently triggering locals saving in the repl for every element
in scope.manglings, but unfortunately currently `global` adds to this
by way of `compiler.destructure`. We can distinguish between globals
and locals by checking the scope.symmeta table. We only save the
mangled scope key if it's found in symmeta but also not a gensym.