make sed invocations work on both gnu and bsd
sed -i in the gnu version of sed has a shorthand allowing omission of
the temp file extension, which is not supported by the bsd version of
sed. bsd sed is the default on macos, and even macos users with gnu
coreutils installed might still be running bsd sed in scripts by
default. this version of the sed invocation works on both versions of
sed.