Make lambda use error instead of assert; trim one level.
This makes it so that the line number reported by the lambda arglist
check (but *not* the traceback) should line up to the caller rather
than the site of the check.

However, currently fennel.traceback strips out this line information,
because most of the time it's redundant. We'll need further work to
stop doing that munging. But for errors in AOT or errors handled by
pcall, this will help make that line number more useful.