weaken the fuzzer
At some point in the past, I added an invariant to the fuzzer to assert
that if Fennel emits code, it should be valid Lua code. In theory this
is a good invariant, but in practice there are forms such as
(pick-values 99999), or bit operations in lua 5.1, where Fennel outputs
code that Lua doesn't compile, but that don't indicate a bug. Because
these fuzz tests run in CI, it would be better to remove this check.