Restrict mode=t in fennel.load-code for luajit.
In Lua 5.2+ the `load' function accepts both bytecode and text.
Unfortunately it *defaults* to accepting either one, but you can fix
this by explicitly setting mode=t for text only.

PUC Lua 5.1 has `loadstring' instead of `load' and it does not have
this safety feature, but LuaJIT's `loadstring' does; it will refuse to
load bytecode when mode=t.

PUC 5.1 will simply ignore the mode argument and happily load the
unsafe bytecode. =(

This matters because most of the sandbox escapes that have been
discovered so far in the Lua VM require loading malicious bytecode.
Lua's default behavior of allowing loading bytecode when mode is
unspecified is a terrible mistake; hopefully it is fixed in future
versions, but for the time being we can at least make LuaJIT a little
safer for calls that go thru Fennel.