Support reading from files in compiler scope.
The macro sandbox is helpful in that it allows untrusted code to be
loaded and analyzed, but the thing that really makes it safe is that
it does not allow *output*.

There are some legitimately safe patterns that we currently don't
allow without disabling the macro sandbox entirely, such as
hugsql-style database access. It would be safer if we allowed these
to run without disabling the macro sandbox.