Add :extra-compiler-env option.
Without this you can provide :compiler-env but then you're on your own
to include all the basics like pairs/ipairs/table; it's very easy to
accidentally leave something out or include something that turns out
to be unsafe.

This also makes it more consistent with the :extra-env option that
works for the runtime environment.