launcher: format helptext like gnu utils
Tools like zsh's _gnu_generic completion helper attempt to infer command
line flags from the --help output.

While this isn't exactly a standard, putting the short opts before the
long ones with a comma separator allows `compdef _gnu_generic fennel` to
configure complain on both the short and long versions of CLI flags,
which is a very handy improvement for a small formatting change!

The bash equivalent would be `complete -F _longopt`, which only allows
completion on the long options with both the new and old help formats.

There is also `_parse_help` and `_parse_usage` in bash, but with or
without this change, it seems to cause completion to freeze. Ideally the
help text would support both, so this may bear further tinkering, but
this is a small step in the right direction.