Use root scope when emitting included fennel
The lambdaAsFn option wasn't propagating during AOT compilation of
required modules when using requireAsInclude. The top-level module's
compilation would have the option applied, but not the required modules.
With the code previously using the utils.root.scope.parent, it would seem
this skipped over the scope established in compile-asts and goes straight
to the global scope. I suspect this propagation issue also impacted the
toBeClosed and assertAsRepl settings, but haven't tested those to be
sure.