Fix bug with ... on case; add test.
    (case ... {: a} a _ 0)

In 1.6.0 this expanded to:

    if ((_G.type(...) == "table") and (nil ~= (...).a)) then
      local a = (...).a
      return a
    else
      local _ = ...
      return 0
    end

This would cause a runtime error when ... was zero values.

This broke because of a fix to the double-eval-safe? function which
caused it to (correctly) say that varg is double-eval safe. The
problem is that we were unknowingly using that function to protect
against BOTH double-evals AND zero-values. By treating varg as
double-eval safe we stopped binding it to a local, which allowed it to
flow straight to _G.type and explode.

The fix is to add an explicit check for varg? when we decide whether
to let-bind the expression being matched against. So now the same code
expands to:

    local case_1_ = ...
    if ((_G.type(case_1_) == "table") and (nil ~= case_1_.a)) then
      local a = case_1_.a
      return a
    else
      local _ = case_1_
      return 0
    end

When ... is zero-valued, case_1_ becomes nil, protecting the call to
_G.type from multivalued nonsense.